* Switch file pointer offsets to longs instead of ints.

+++ trunk/jdem846/src/us/wthr/jdem846/input/esri/GridAsciiDataCache.java
@@ -115,7 +115,7 @@ public class GridAsciiDataCache extends DataCache
 		Arrays.fill(load_buffer, (byte)0x0);
 		
 
-		int seekStart = start * (Float.SIZE / 8);
+		long seekStart = (long)start * (long)  (Float.SIZE / 8);
 		try {
 			inputData.seek(seekStart);
 		} catch (IOException ex) {
